requirements.txt
Pillow>=5.3.0
pycodestyle>=2.4.0
pyflakes>=2.0.0
